Move UT file location
