+++ package.json
@@ -12,7 +12,7 @@
     "build": "run-p compile-external-types compile-v3-contract-types && tsc",
     "clean": "rm -rf dist cdk.out",
     "watch": "tsc -w",
-    "test:unit": "jest test/jest/unit/ && ts-mocha -p tsconfig.cdk.json lib/rpc/**/*.test.ts",
+    "test:unit": "jest test/jest/unit/ && ts-mocha -p tsconfig.cdk.json test/mocha/unit/**/*",
     "test:unit:watch": "jest test/jest/unit/ --watch",
     "test:integ": "ts-mocha -p tsconfig.cdk.json -r test/mocha/dynamoDBLocalFixture.ts test/mocha/integ --recursive --extension .test.ts",
     "test:e2e": "ts-mocha -p tsconfig.cdk.json -r dotenv/config test/mocha/e2e/**/*.test.ts",
